requirements.txt
# === Core ML / AI (CPU wheels only) ===
torch==2.3.0
torchaudio==2.3.0
transformers==4.45.1

# === Speech & Text (prebuilt wheels) ===
TTS==0.22.0
openai-whisper==20231117

# === NLP / Config ===
spacy==3.7.5
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

# === Audio Processing ===
pydub==0.25.1
soundfile==0.12.1

# === Utilities & Interface ===
tqdm==4.66.5
numpy==1.22.0
scipy==1.11.2
pandas==1.5.3
scikit-learn==1.5.1
rich==13.7.1
gradio==4.44.0
